move filter logic to layer buttons
